+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Optimized the conversion of Rhino Meshes to COMPAS meshes.
+
 ### Removed